requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.13
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.8.6
psycopg2-binary==2.8.5
pycodestyle==2.8.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0